Includes new files to hold teh decoupled ECM. Includes full initalization migrated over tothe new decoupled implementation.:
