+++ config/PhysiCell_settings.xml
@@ -19,7 +19,7 @@
 	</overall>
 	
 	<parallel>
-		<omp_num_threads>1</omp_num_threads>
+		<omp_num_threads>8</omp_num_threads>
 	</parallel> 
 	
 	<save>
@@ -74,6 +74,11 @@
 
 		<Microenvironment_parameters description="Microenvironment parameters" type="double" units="dimensionlesss">0</Microenvironment_parameters>
 		<chemotactic_substrate_decay_rate description="Chemotactic subsrate decay rate" type="double" units="1/minutes">0.1</chemotactic_substrate_decay_rate>
+
+		<ECM_unit_size description="Size of ECM units" type="double" units="dimensionlesss">0</ECM_unit_size>
+		<ECM_dx description="Length of ECM unit in x-direction" type="double" units="um">80</ECM_dx>
+		<ECM_dy description="Length of ECM unit in y-direction" type="double" units="um">80</ECM_dy>
+		<ECM_dz description="Length of ECM unit in z-direction" type="double" units="um">80</ECM_dz>
 
 		<Model_Selection_Parameters description="Allows selection of various cell motility-ECM models" type="double">0</Model_Selection_Parameters>
 		<cell_motility_ECM_interaction_model_selector description="Specifies the cell-ECM interaction model: follower chemotaxis/no follower hysteresis, follower hysteresis/no follower chemotaxis" type="string">follower hysteresis/no follower chemotaxis</cell_motility_ECM_interaction_model_selector>
